edit calc.js to have floats instead of ints

+++ js/calc.js
@@ -1,9 +1,9 @@
 $(document).ready(function() {
 
 	$("#count").on("click", function () {
-		var alc = parseInt($("#alc").val());
-		var cap = parseInt($("#cap").val());
-		var price = parseInt($("#price").val());
+		var alc = parseFloat($("#alc").val());
+		var cap = parseFloat($("#cap").val());
+		var price = parseFloat($("#price").val());
 		
 		$("#result").val((cap*(alc/100)/price).toFixed(2));
 	});
